# hospital-app-service/src/main/resources/schema/hospital-api-docs.yaml
openapi: 3.0.1
info:
  title: Tutorial Management API
  description: This API exposes endpoints to manage tutorials.
  termsOfService: https://www.hospital.com/terms
  contact:
    name: Amachi
    url: https://www.hospital.com
    email: hospital@gmail.com
  license:
    name: MIT License
    url: https://choosealicense.com/licenses/mit/
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Server URL in Development environment
- url: https://hospital-api.com
  description: Server URL in Production environment
...
paths:
...
  /medical/addresses/{id}:
    get:
      tags:
      - Address
      summary: Retourne un Address
      description: Envoye en tant que query parameters
      operationId: getAddress
      parameters:
      - name: id
        in: path
        description: Return objet Address
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "401":
          description: Non autorise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "405":
          description: Entrée invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "403":
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "200":
          description: Get Address appelee avec succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "404":
          description: Page non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "500":
          description: Le serveur n'a pas répondu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
    put:
      tags:
      - Address
      summary: Modifie un Address
      description: Envoye en tant que body parameters
      operationId: updateAddress
      parameters:
      - name: id
        in: path
        description: Modifie objet Address
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressDTO'
        required: true
      responses:
        "200":
          description: Address modifie avec succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "401":
          description: Non autorise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "403":
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "404":
          description: Page non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "405":
          description: Entrée invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "500":
          description: Le serveur n'a pas répondu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
    delete:
      tags:
      - Address
      summary: Supprime un Address
      description: Envoye en tant que query parameters
      operationId: deleteAddress
      parameters:
      - name: id
        in: path
        description: Delete objet Address
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Delete Address appelée avec succes
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Non autorise
          content:
            application/json:
              schema:
                type: object
        "403":
          description: Accès refusé
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Page non trouvée
          content:
            application/json:
              schema:
                type: object
        "405":
          description: Entrée invalide
          content:
            application/json:
              schema:
                type: object
        "500":
          description: Le serveur n'a pas répondu
          content:
            application/json:
              schema:
                type: object
...
  /medical/addresses:
    get:
      tags:
      - Address
      summary: Retourne la liste de Addresses
      description: Envoye en tant que path variables json format
      operationId: getAddresses
      parameters:
      - name: addressSearchDTO
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/AddressSearchDTO'
      - name: pageNumber
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: sort
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: List de Address retournée avec succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAddressDTO'
        "401":
          description: Non autorise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAddressDTO'
        "403":
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAddressDTO'
        "404":
          description: Page non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAddressDTO'
        "405":
          description: Entrée invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAddressDTO'
        "500":
          description: Le serveur n'a pas répondu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAddressDTO'
    post:
      tags:
      - Address
      summary: Cree un Address
      description: Envoye en tant que body parameters
      operationId: addAddress
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressDTO'
        required: true
      responses:
        "200":
          description: Address cree avec succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "401":
          description: Non autorise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "403":
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "404":
          description: Page non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "405":
          description: Entrée invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
        "500":
          description: Le serveur n'a pas répondu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDTO'
...
  /medical/addressesallparams:
    get:
      tags:
      - Address
      summary: Retourne la liste de Addresses
      description: Envoye en tant que path variables json format
      operationId: findAllAddresses
      responses:
        "200":
          description: List de Address retournée avec succes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDTO'
        "401":
          description: Non autorise
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDTO'
        "403":
          description: Accès refusé
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDTO'
        "404":
          description: Page non trouvée
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDTO'
        "405":
          description: Entrée invalide
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDTO'
        "500":
          description: Le serveur n'a pas répondu
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDTO'
components:
  schemas:
    CountryDTO:
      required:
      - codeCountry
      type: object
      properties:
        idCountry:
          type: integer
          format: int32
        codeCountry:
          type: string
        name:
          maxLength: 255
          minLength: 0
          type: string
        codePhone:
          type: integer
          format: int32
        codeAlpha3:
          type: string
        codeNumeric:
          type: integer
          format: int32
        nationality:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
      description: Créé objet Country
    AddressDTO:
      type: object
      properties:
        idAddress:
          type: integer
          format: int64
        address1:
          type: string
        city:
          type: string
        address2:
          type: string
        postalCode:
          type: string
        location:
          type: string
        country:
          $ref: '#/components/schemas/CountryDTO'
      description: Créé objet Address
...
    AddressSearchDTO:
      type: object
      properties:
        idAddress:
          type: integer
          format: int64
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
    PageAddressDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/AddressDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        unpaged:
          type: boolean
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean